+++ src/utils/customChains.ts
@@ -55,32 +55,3 @@ export const hyperEVM = /*#__PURE__*/ defineChain({
     },
     testnet: false,
 });
-
-export const plasma = /*#__PURE__*/ defineChain({
-    id: 9745,
-    name: 'Plasma',
-    nativeCurrency: {
-        name: 'XPL',
-        symbol: 'XPL',
-        decimals: 18,
-    },
-    rpcUrls: {
-        default: {
-            http: ['https://rpc.plasma.to'],
-            webSocket: ['wss://rpc.plasma.to'],
-        },
-    },
-    blockExplorers: {
-        default: {
-            name: 'Plasma Explorer',
-            url: 'https://plasmascan.to',
-            apiUrl: 'https://api.routescan.io/v2/network/mainnet/evm/9745/etherscan/api?',
-        },
-    },
-    contracts: {
-        multicall3: {
-            address: '0xcA11bde05977b3631167028862bE2a173976CA11',
-            blockCreated: 1,
-        },
-    },
-});